refactor: move routing files to proxy/routing/ subdirectory
- mapping-resolver, model-state, overflow, usage-window-tracker, enhancement-config -> proxy/routing/
- Fixed all import paths in proxy/, enhancement/, admin/, src/index.ts, tests/
- All 580 tests pass
