+++ src/index.ts
@@ -23,11 +23,11 @@ import { adminRoutes } from "./admin/routes.js";
 import { RetryRuleMatcher } from "./proxy/orchestration/retry-rules.js";
 import { ProviderSemaphoreManager } from "./proxy/orchestration/semaphore.js";
 import { AdaptiveConcurrencyController } from "./proxy/adaptive-controller.js";
-import { loadEnhancementConfig } from "./proxy/enhancement-config.js";
+import { loadEnhancementConfig } from "./proxy/routing/enhancement-config.js";
 import type { StateRegistry } from "./core/registry.js";
 import { RequestTracker } from "./monitor/request-tracker.js";
-import { modelState } from "./proxy/model-state.js";
-import { UsageWindowTracker } from "./proxy/usage-window-tracker.js";
+import { modelState } from "./proxy/routing/model-state.js";
+import { UsageWindowTracker } from "./proxy/routing/usage-window-tracker.js";
 import { SessionTracker } from "./proxy/loop-prevention/session-tracker.js";
 import { DEFAULT_LOOP_PREVENTION_CONFIG } from "./proxy/loop-prevention/types.js";
 import { scheduleLogCleanup } from "./db/log-cleaner.js";
