+++ src/proxy/openai.ts
@@ -13,7 +13,7 @@ import { createOrchestrator } from "./orchestration/orchestrator.js";
 import { RetryRuleMatcher } from "./orchestration/retry-rules.js";
 import { ProviderSemaphoreManager } from "./orchestration/semaphore.js";
 import type { RequestTracker } from "../monitor/request-tracker.js";
-import type { UsageWindowTracker } from "./usage-window-tracker.js";
+import type { UsageWindowTracker } from "./routing/usage-window-tracker.js";
 import type { AdaptiveConcurrencyController } from "./adaptive-controller.js";
 import { HTTP_NOT_FOUND, HTTP_BAD_GATEWAY } from "../core/constants.js";
